src/requirements.txt
alembic==1.16.1
aniso8601==10.0.1
APScheduler==3.11.0
attrs==25.3.0
blinker==1.9.0
click==8.2.1
dnspython==2.7.0
email-validator==1.3.1
Flask==2.3.3
Flask-APScheduler==1.13.1
Flask-Login==0.6.2
Flask-Migrate==4.0.5
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
greenlet==3.2.3
gunicorn==21.2.0
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.3
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
pytz==2025.2
referencing==0.36.2
rpds-py==0.25.1
six==1.17.0
SQLAlchemy==2.0.30
tomli==2.2.1
typing_extensions==4.14.0
tzlocal==5.3.1
Werkzeug==2.3.8
WTForms==3.1.2
